stress-test/requirements.txt
# Stress testing dependencies
locust>=2.15.0
aiohttp>=3.9.0
pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0
